extracted StopTimer class
