+++ .idea/.name
@@ -1,1 +1,1 @@
-ShotTimer
+BengShotTimer
