+++ .idea/modules.xml
@@ -2,7 +2,7 @@
 <project version="4">
   <component name="ProjectModuleManager">
     <modules>
-      <module fileurl="file://$PROJECT_DIR$/ShotTimerForked.iml" filepath="$PROJECT_DIR$/ShotTimerForked.iml" />
+      <module fileurl="file://$PROJECT_DIR$/BengShotTimer.iml" filepath="$PROJECT_DIR$/BengShotTimer.iml" />
       <module fileurl="file://$PROJECT_DIR$/app/app.iml" filepath="$PROJECT_DIR$/app/app.iml" />
     </modules>
   </component>
